requirements.txt
pandas>=1.5.0
numpy>=1.23.0
statsmodels>=0.14.0
scipy>=1.10.0
scikit-learn>=1.2.0  
matplotlib>=3.7.0
psycopg2>=2.9.0
optuna>=3.0.0
tabulate>=0.9.0
plotly>=5.22.0
python-dotenv
pymc
pykalman>=0.9.5